code/requirements.txt
paho-mqtt
kafka-python
influxdb